+++ tests/testthat/test-golem.R
@@ -71,4 +71,29 @@ test_that("Golem no CDN", {
   expect_message(bundle())
   setwd(wd)
   delete_tmp_package(pkg)
+  
+  # test framework7
+  pkg <- create_tmp_golem()
+  setwd(pkg)
+  expect_output(scaffold_golem(framework7 = TRUE, edit = FALSE))
+  expect_message(bundle())
+  setwd(wd)
+  delete_tmp_package(pkg)
+})
+
+test_that("Golem F7", {
+
+  # keep working directory
+  wd <- getwd()
+
+  # test bare
+  pkg <- create_tmp_golem()
+  setwd(pkg)
+  on.exit({
+    setwd(wd)
+    delete_tmp_package(pkg)
+  })
+  expect_output(scaffold_golem(framework7 = TRUE, edit = FALSE))
+  expect_error(scaffold_golem(edit = FALSE))
+  expect_message(bundle_dev())
 })